feat: tree

+++ src/main/java/leetcode/problems/LeetCode319.java
@@ -8,7 +8,7 @@ public class LeetCode319 {
 
     public int bulbSwitch(int n) {
 
-        return (int) (Math.sqrt((double) n));
+        return (int) (Math.sqrt(n));
     }
 
 }
